cloud_run_function_gcp/requirements-gcp.txt
functions-framework==3.*
requests
pandas
pyarrow
google-cloud-storage
